==> 15_ML_RM_Fuel_Efficieny/03_Notebooks/02_Desarrollo/01_Set Up.ipynb <==
# SET UP

## IMPORTAR PAQUETES

import pandas as pd

## CREAR LOS DATASETS INICIALES

### IMPORTAR LOS DATOS

df = pd.read_csv('../../02_Datos/01_Originales/auto-mpg.data',sep='\s+', decimal= '.',names=['mpg','cilindros','desplazamiento','caballos','peso','aceleracion','any_modelo','origen','nombre_auto'])
df

### EXTRAER Y RESERVAR EL DATASET DE VALIDACIÓN

val = df.sample(frac= 0.3)

#### Guardamos dataset de validación

val.to_csv('../../02_Datos/02_Validacion/validacion.csv')

### EXTRAER Y GUARDAR EL DATASET DE TRABAJO

trabajo = df.loc[~df.index.isin(val.index)]

#### Guardamos dataset de trabajo

trabajo.to_csv('../../02_Datos/03_Trabajo/trabajo.csv')
