==> 15_ML_RM_Fuel_Efficieny/03_Notebooks/02_Desarrollo/04_Transformación de los datos.ipynb <==
# 4 - TRANSFORMACIÓN DE DATOS

## IMPORTAR PAQUETES

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline

from sklearn.preprocessing import OneHotEncoder
from sklearn.preprocessing import OrdinalEncoder
from sklearn.preprocessing import PowerTransformer
from sklearn.preprocessing import MinMaxScaler
from sklearn.preprocessing import StandardScaler

## IMPORTAR LOS DATOS

df = pd.read_pickle('../../02_Datos/03_Trabajo/tablon_analitico.pickle')
df.head()

x = df.drop(columns='mpg').copy()
target = df.mpg.copy().reset_index(drop=True)


## TRANSFORMACIÓN DE CATEGÓRICAS

### ONE HOT ENCODER

#### Variable a aplicar OHE

var_ohe = ['any_modelo','origen']

#### Instanciar

ohe = OneHotEncoder(sparse_output= False, handle_unknown= 'ignore')

#### Entrenar y aplicar

cat_ohe = ohe.fit_transform(x[var_ohe])

cat_ohe = pd.DataFrame(cat_ohe, columns = ohe.get_feature_names_out())

#Eliminamos variables

x = x.drop(columns= x[var_ohe])

#### Guardar como dataframe

### ORDINAL ENCODING

df.cilindros.value_counts()

var_oe = ['cilindros']

#Orden de la primera variable
orden_formacion = [2,3,4,5,6,8]

oe = OrdinalEncoder(categories = [orden_formacion],
                    handle_unknown = 'use_encoded_value',
                    unknown_value = -1)

cat_oe = oe.fit_transform(x[var_oe])

#Añadir sufijos a los nombres
nombres_oe = [variable + '_oe' for variable in var_oe]

#Guardar como dataframe
cat_oe = pd.DataFrame(cat_oe, columns = nombres_oe)


Quitamos la variable cilindros del dataframe

x = x.drop(columns='cilindros')

x = pd.concat([x,cat_ohe,cat_oe],axis= 1)

## TRANSFORMACIÓN DE VARIABLES NUMÉRICAS

### CON YEO-JOHNSON

#### Variables a normalizar con Yeo_Johnson

var_yeo = ['desplazamiento','caballos','peso','aceleracion']

#### Instanciar 

yeo = PowerTransformer(method='yeo-johnson')

#### Entrenar y aplicar

num_yeo = yeo.fit_transform(x[var_yeo])

#### Guardar como dataframe

#Añadir sufijos a los nombres
nombres_yeo = [variable + '_yeo' for variable in var_yeo]

#Guardar como dataframe
num_yeo = pd.DataFrame(num_yeo,columns= nombres_yeo)

#### Eliminar y unificar variables

#### Eliminar variables originales
x.drop(columns= var_yeo,inplace= True)

x = pd.concat([x, num_yeo], axis=1)

### RESCALAR VARIABLES

#### Variables a reescalar con Min-Max

var_mms = x.iloc[:,89:].columns

#### Instanciar

mms = MinMaxScaler()

#### Entrenar y aplicar

df_mms = mms.fit_transform(x[var_mms])

#### Guardar como Dataframe

#Añadir sufijos a los nombres

nombres_mms = [variable + '_mms' for variable in var_mms]

df_mms = pd.DataFrame(df_mms, columns= nombres_mms)

#### Eliminar y unificar variables

x.drop(columns= var_mms, inplace= True)

x = pd.concat([x,df_mms,target],axis=1)

## GUARDAR DATASET TRAS TRANSFORMACIÓN DE DATOS

x.to_pickle('../../02_Datos/03_Trabajo/df_tablon.pickle')
