==> 15_ML_RM_Fuel_Efficieny/03_Notebooks/02_Desarrollo/08_Preselección_Modelos_LayzPredict.ipynb <==
# 8 - PRESELECCIÓN DEL MODELO LAZY PREDICT

## IMPORTACIÓN DE PAQUETES

import numpy as np 
import pandas as pd

from lazypredict.Supervised import LazyRegressor
from sklearn.model_selection import train_test_split

## CARGAR LOS DATOS

df = pd.read_pickle('../../02_Datos/03_Trabajo/df_final.pickle')
df.head()

## SEPARAR PREDICTORAS Y TARGET

x = df.drop(columns= 'mpg').copy()
y = df.mpg.copy()

x.head()

y.value_counts()

## SEPARAR DATASET DE ENTRENAMIENTO Y VALIDACIÓN

train_x, val_x, train_y, val_y = train_test_split(x, y, test_size= 0.3, random_state= 42)

## MODELIZAR SEGÚN LAZYPREDICT

reg = LazyRegressor(random_state= 42, verbose=0)
models, predictions = reg.fit(train_x, val_x, train_y, val_y)

models

## CONCLUSIÓN

**PRÓXIMOS PASOS:** 

1. Realizar predicción con modelos bases (sin hiperparámetros) y incluiremos el XGBoost.
2. Preselección de variables según modelo base.
3. Realizar predicción con modelo + hiperparámetros + preselección de variables.
4. Evaluar métricas
